Project/BotDeeptry123/requirements.txt
 
aenum==2.1.2
attrs==18.2.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
fbchat==1.6.4
idna==2.8
requests==2.21.0
soupsieve==1.8
urllib3==1.24.2
python-aiml==0.9.3
google-api-python-client==1.8.0
python-http-client==3.2.7
